requirements.txt
Flask>=2.3,<4.0
Flask-Login>=0.6,<0.7
Flask-WTF>=1.1,<2.0
Flask-SQLAlchemy>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0

# Opzionale: librerie per importare file Excel/CSV nelle distinte.
# Pandas has been removed to avoid heavy build-time dependencies on Windows.  
# Instead, the import feature uses the built-in csv module for CSV files and
# the openpyxl package for .xlsx files.  
openpyxl>=3.1
